+++ src/tterm_helper.ml
@@ -44,6 +44,9 @@ let rec t_free_vars t =
           Svs.empty pl
       in
       Svs.union t_fvs pl_fvs
+  | Tlambda (ps, t) ->
+      let t_fvs = t_free_vars t in
+      List.fold_right (fun p fvs -> Svs.diff fvs (p_vars p)) ps t_fvs
   | Tquant (_, vl, t) -> Svs.diff (t_free_vars t) (Svs.of_list vl)
   | Tbinop (_, t1, t2) -> Svs.union (t_free_vars t1) (t_free_vars t2)
   | Tnot t -> t_free_vars t
@@ -76,30 +79,45 @@ let t_ty_check t ty =
   | None, None -> ()
 
 let ls_arg_inst ls tl =
-  try
-    List.fold_left2
-      (fun tvm ty t -> ty_match tvm ty (t_type t))
-      Mtv.empty ls.ls_args tl
-  with Invalid_argument _ ->
-    let loc = (List.hd tl).t_loc in
-    W.error ~loc
-      (W.Bad_arity (ls.ls_name.id_str, List.length ls.ls_args, List.length tl))
+  let rec short_fold_left2 f accu l1 l2 =
+    match (l1, l2) with
+    | a1 :: l1, a2 :: l2 -> short_fold_left2 f (f accu a1 a2) l1 l2
+    | _, _ -> accu
+  in
+  short_fold_left2
+    (fun tvm ty t -> ty_match tvm ty (t_type t))
+    Mtv.empty ls.ls_args tl
 
-let ls_app_inst ls tl ty =
+let drop n xs =
+  let rec aux n xs =
+    match (n, xs) with
+    | 0, xs -> xs
+    | _, [] -> []
+    | n, _ :: xs -> aux (n - 1) xs
+  in
+  if n < 0 then invalid_arg "drop" else aux n xs
+
+let ls_app_inst ls tl ty loc =
   let s = ls_arg_inst ls tl in
   match (ls.ls_value, ty) with
-  | Some _, None ->
-      (* FIXME: get a proper location here *)
-      W.error ~loc:Location.none (W.Predicate_symbol_expected ls.ls_name.id_str)
-  | None, Some _ ->
-      (* FIXME: get a proper location here *)
-      W.error ~loc:Location.none (W.Function_symbol_expected ls.ls_name.id_str)
-  | Some vty, Some ty -> ty_match s vty ty
+  | Some _, None -> W.error ~loc (W.Predicate_symbol_expected ls.ls_name.id_str)
+  | None, Some _ -> W.error ~loc (W.Function_symbol_expected ls.ls_name.id_str)
+  | Some vty, Some ty ->
+      let vty =
+        let ntl = List.length tl in
+        if ntl >= List.length ls.ls_args then vty
+        else
+          (* build the result type in case of a partial application *)
+          List.fold_right
+            (fun t1 t2 -> { ty_node = Tyapp (ts_arrow, [ t1; t2 ]) })
+            (drop ntl ls.ls_args) vty
+      in
+      ty_match s vty ty
   | None, None -> s
 
 (** Pattern constructors *)
 
-let mk_pattern p_node p_ty = { p_node; p_ty; p_loc = None }
+let mk_pattern p_node p_ty p_loc = { p_node; p_ty; p_loc }
 let p_wild ty = mk_pattern Pwild ty
 let p_var vs = mk_pattern (Pvar vs) vs.vs_ty
 let p_app ls pl ty = mk_pattern (Papp (ls, pl)) ty
@@ -126,13 +144,13 @@ let mk_term t_node t_ty t_loc = { t_node; t_ty; t_attrs = []; t_loc }
 let t_var vs = mk_term (Tvar vs) (Some vs.vs_ty)
 let t_const c ty = mk_term (Tconst c) (Some ty)
 
-let t_app ls tl ty =
-  ignore (ls_app_inst ls tl ty : ty Mtv.t);
-  mk_term (Tapp (ls, tl)) ty
+let t_app ls tl ty loc =
+  ignore (ls_app_inst ls tl ty loc : ty Mtv.t);
+  mk_term (Tapp (ls, tl)) ty loc
 
-let t_field t ls ty =
-  ignore (ls_app_inst ls [ t ] ty : ty Mtv.t);
-  mk_term (Tfield (t, ls)) ty
+let t_field t ls ty loc =
+  ignore (ls_app_inst ls [ t ] ty loc : ty Mtv.t);
+  mk_term (Tfield (t, ls)) ty loc
 
 let t_if t1 t2 t3 = mk_term (Tif (t1, t2, t3)) t2.t_ty
 let t_let vs t1 t2 = mk_term (Tlet (vs, t1, t2)) t2.t_ty
@@ -143,6 +161,7 @@ let t_case t1 ptl =
   | (_, _, t) :: _ -> mk_term (Tcase (t1, ptl)) t.t_ty
 
 let t_quant q vsl t ty = mk_term (Tquant (q, vsl, t)) ty
+let t_lambda ps t ty = mk_term (Tlambda (ps, t)) ty
 let t_binop b t1 t2 = mk_term (Tbinop (b, t1, t2)) None
 let t_not t = mk_term (Tnot t) None
 let t_old t = mk_term (Told t) t.t_ty
@@ -157,17 +176,14 @@ let f_binop op f1 f2 = t_binop op (t_prop f1) (t_prop f2)
 let f_not f = t_not (t_prop f)
 
 let t_quant q vsl t ty loc =
-  match (q, vsl) with
-  | Tlambda, [] -> t
-  | _, [] -> t_prop t
-  | Tlambda, _ -> t_quant q vsl t ty loc
-  | _, _ ->
+  match vsl with
+  | [] -> t_prop t
+  | _ ->
       if ty <> None then W.error ~loc W.Formula_expected;
       t_quant q vsl (t_prop t) None loc
 
 let f_forall = t_quant Tforall
 let f_exists = t_quant Texists
-let t_lambda = t_quant Tlambda
 let f_and = f_binop Tand
 let f_and_asym = f_binop Tand_asym
 let f_or = f_binop Tor