+++ test/patterns/fun_pair.mli
@@ -1,0 +1,8 @@
+val f : (int * 'a) list -> int list
+(*@ ys = f xs
+    ensures ys = List.map (fun (x, _) -> x) xs
+*)
+
+(* {gospel_expected|
+   [0] OK
+   |gospel_expected} *)